// src/platform/plugins/shared/workflows_extensions/server/steps/data/data_find_step.d.ts
export declare const dataFindStepDefinition: import("../../step_registry/types").ServerStepDefinition<import("zod").ZodObject<{
    condition: import("zod").ZodString;
    errorIfEmpty: import("zod").ZodOptional<import("zod").ZodBoolean>;
}, import("zod/v4/core").$strip>, import("zod").ZodObject<{
    item: import("zod").ZodNullable<import("zod").ZodUnknown>;
    index: import("zod").ZodNullable<import("zod").ZodNumber>;
}, import("zod/v4/core").$strip>, import("zod").ZodObject<{
    items: import("zod").ZodUnknown;
}, import("zod/v4/core").$strip>>;